Add prepare/exec/query/finalize integration tests

+++ c/src/request.h
@@ -35,7 +35,7 @@
 
 #define DQLITE__REQUEST_SCHEMA_FINALIZE(X, ...)	\
 	X(uint32, db_id, __VA_ARGS__)		\
-	X(uint64, stmt_id, __VA_ARGS__)
+	X(uint32, stmt_id, __VA_ARGS__)
 
 DQLITE__SCHEMA_DEFINE(dqlite__request_helo,      DQLITE__REQUEST_SCHEMA_HELO);
 DQLITE__SCHEMA_DEFINE(dqlite__request_heartbeat, DQLITE__REQUEST_SCHEMA_HEARTBEAT);
